ardk UPM 3.4.0

+++ Runtime/PlatformAdapterManager/AbstractTexturesSetter.cs
@@ -28,7 +28,7 @@ namespace Niantic.Lightship.AR.PAM
         public virtual double GetCurrentTimestampMs()
         {
             if (PlatformDataAcquirer.TryGetCameraFrame(out XRCameraFrame frame))
-                return frame.timestampNs / MillisecondToNanosecondFactor;
+                return (double)frame.timestampNs / MillisecondToNanosecondFactor;
 
             return 0;
         }
